+++ config/auth.php
@@ -106,6 +106,7 @@ return [
     'socialite' => [
         'drivers' => [
             'google',
+            'facebook',
         ],
     ],